perf(subconscious): eliminate resource waste in delta indexing (v0.21.24)
5 optimizations that dramatically reduce memory, CPU, and TCP overhead:

1. File-size delta detection: Use fs.statSync().size to skip unchanged files
   without reading them. Previously every cycle read entire .jsonl files
   just to count lines.

2. Direct function call: maintainMemory() now calls runIndexDelta() directly
   instead of HTTP self-fetch to localhost. Eliminates TCP connection,
   HTTP parsing, and JSON serialization per agent per cycle.

3. Global project discovery cache: ~/.claude/projects/ scan results cached
   with 10-min TTL. Previously each of ~108 agents independently scanned
   and read the first 20 lines of every .jsonl file.

4. Single-read pipeline: Conversations needing indexing are read once.
   Previously each file was read 3 times (filter, dry-run, index).

5. Extracted lib/index-delta.ts: Core logic extracted from API route into
   a module callable both directly and via HTTP. API route is now a thin
   wrapper.
