+++ data/help-embeddings.json
@@ -1,6 +1,6 @@
 {
   "modelVersion": "Xenova/bge-small-en-v1.5",
-  "generatedAt": "2026-02-08T04:34:39.677Z",
+  "generatedAt": "2026-02-08T05:31:31.293Z",
   "documentCount": 136,
   "documents": [
     {
