+++ version.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.21.23",
+  "version": "0.21.24",
   "releaseDate": "2026-02-07",
   "changelog": "https://github.com/23blocks-OS/ai-maestro/releases",
   "minSupportedVersion": "0.10.0"
